refactor: create user icon drop down component
