+++ src/app/components/header/header.tsx
@@ -1,14 +1,10 @@
-import React, { FC, Fragment } from 'react';
+import React, { FC } from 'react';
 import { Link } from 'react-router-dom';
-import { Menu, Transition } from '@headlessui/react';
-import { Bars3Icon, UserIcon } from '@heroicons/react/24/outline';
+import { Bars3Icon } from '@heroicons/react/24/outline';
 import mocks from '../mocks-data/mocks';
 import { useStateContext } from '../../state/state-context';
 import CartIcon from './cart-icon';
-
-function classNames(...classes: string[]) {
-  return classes.filter(Boolean).join(' ');
-}
+import UserIconDropdown from './user-icon-dropdown';
 
 interface HeaderProps {
   open: boolean;
@@ -44,7 +40,7 @@ const Header: FC<HeaderProps> = ({ setOpen }) => {
                 </div>
               </div>
 
-              {/* Mobile menu and search (lg-) */}
+              {/* Mobile menu (lg-) */}
               <div className="flex flex-1 items-center lg:hidden">
                 <button
                   type="button"
@@ -65,62 +61,8 @@ const Header: FC<HeaderProps> = ({ setOpen }) => {
               <div className="flex flex-1 items-center justify-end">
                 {auth.get.isAuth ? (
                   <div className="flex items-center lg:ml-8">
-                    {/* <button onClick={auth.logout} className="text-sm font-medium text-gray-500 hover:text-gray-700"> */}
-                    {/*  Logout */}
-                    {/* </button> */}
-                    {/* Account */}
-                    <div className="flex">
-                      <Menu as="div" className="relative ml-3">
-                        <div>
-                          <Menu.Button className="relative flex -m-2 p-2 text-sm text-gray-400 hover:text-gray-500 rounded-full">
-                            <span className="sr-only">Open user menu</span>
-                            <UserIcon className="h-6 w-6" aria-hidden="true" />
-                          </Menu.Button>
-                        </div>
-                        <Transition
-                          as={Fragment}
-                          enter="transition ease-out duration-100"
-                          enterFrom="transform opacity-0 scale-95"
-                          enterTo="transform opacity-100 scale-100"
-                          leave="transition ease-in duration-75"
-                          leaveFrom="transform opacity-100 scale-100"
-                          leaveTo="transform opacity-0 scale-95"
-                        >
-                          <Menu.Items className="absolute right-0 z-10 mt-2 w-48 origin-top-right rounded-md bg-white py-1 shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
-                            <Menu.Item>
-                              {({ active }) => (
-                                <Link
-                                  to="/profile"
-                                  className={classNames(
-                                    active ? 'bg-gray-200' : '',
-                                    'block w-full text-center px-4 py-2 text-sm text-gray-700'
-                                  )}
-                                >
-                                  Your Profile
-                                </Link>
-                              )}
-                            </Menu.Item>
-                            <Menu.Item>
-                              {({ active }) => (
-                                <button
-                                  onClick={auth.logout}
-                                  className={classNames(
-                                    active ? 'bg-gray-200' : '',
-                                    'block w-full px-4 py-2 text-sm text-gray-700'
-                                  )}
-                                >
-                                  Logout
-                                </button>
-                              )}
-                            </Menu.Item>
-                          </Menu.Items>
-                        </Transition>
-                      </Menu>
-                      {/* <Link to="/profile" className="-m-2 p-2 text-gray-400 hover:text-gray-500"> */}
-                      {/*  <span className="sr-only">Account</span> */}
-                      {/*  <UserIcon className="h-6 w-6" aria-hidden="true" /> */}
-                      {/* </Link> */}
-                    </div>
+                    {/* User icon dropdown menu */}
+                    <UserIconDropdown></UserIconDropdown>
 
                     {/* Cart */}
                     <div className="ml-4 flow-root lg:ml-8">
